+++ tests/unit/test.vcxproj
@@ -223,12 +223,14 @@
     <ClCompile Include="ebpfcore_test.cpp" />
     <ClCompile Include="export_program_info_test.cpp" />
     <ClCompile Include="libbpf_test.cpp" />
+    <ClCompile Include="libbpf_test_jit.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\libs\execution_context\unit\execution_context_unit_test_jit.h" />
     <ClInclude Include="..\..\libs\thunk\mock\mock.h" />
     <ClInclude Include="..\end_to_end\helpers.h" />
     <ClInclude Include="..\end_to_end\test_helper.hpp" />
+    <ClInclude Include="libbpf_test_jit.h" />
     <ClCompile Include="wer_report_test_wrapper.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
